Random number generator.
rand.Intn() gives us random integer till the max value we give it as parameter.
But if we use this function barely, it gives us same number everytime we run the program. Basically that's not random anymore. But if we change the random number seeding value everytime, than it'll give us random numbers. Hence, we use the rand.Seed() and to give it a seed value. And to give it a fresh seed value, we give it seed parameters as time.Now() but to give it more accuracy, we give it time.Now().unixNano() which give the program time elapsed from 1 Jan 1970 till right now with nano second accuracy so the seed value changes every nano second.