Basics of Concepts of Functions
A function with the name of startGame() is defined outside of the main(). startGame() contains a variable called 'instructions', which is then used in fmt.Println() but 'instructions' can only be accessed in the startGame() and in turn startGame() can then be accessed in main(). But we can't access 'instructions' in main().
This concept is called as Scope. And that was the main point of this particular program.